# packages/protocol/generated/work-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenBot Work HTTP",
    "version": "1"
  },
  "paths": {
    "/api/v1/tasks": {
      "post": {
        "operationId": "createWorkTask",
        "x-openbot-owner-session": true,
        "x-openbot-allowed-origin": true,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTask"
              }
            }
          }
        },
        "x-openbot-max-body-bytes": 20000,
        "responses": {
          "202": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkSnapshot"
                }
              }
            }
          },
          "401": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "403": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "404": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "408": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "409": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "413": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "422": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "503": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "operationId": "getWorkTask",
        "x-openbot-owner-session": true,
        "x-openbot-allowed-origin": false,
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkSnapshot"
                }
              }
            }
          },
          "401": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "403": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "404": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "408": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "409": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "413": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "422": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "503": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tasks/{task_id}/cancel": {
      "post": {
        "operationId": "cancelWorkTask",
        "x-openbot-owner-session": true,
        "x-openbot-allowed-origin": true,
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyCommand"
              }
            }
          }
        },
        "x-openbot-max-body-bytes": 128,
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkSnapshot"
                }
              }
            }
          },
          "401": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "403": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "404": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "408": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "409": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "413": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "422": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "503": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/corrections": {
      "post": {
        "operationId": "correctWorkTask",
        "x-openbot-owner-session": true,
        "x-openbot-allowed-origin": true,
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCorrection"
              }
            }
          }
        },
        "x-openbot-max-body-bytes": 32768,
        "responses": {
          "202": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkCorrection"
                }
              }
            }
          },
          "401": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "403": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "404": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "408": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "409": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "413": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "422": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "503": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/artifacts/{artifact_id}": {
      "get": {
        "operationId": "downloadWorkArtifact",
        "x-openbot-owner-session": true,
        "x-openbot-allowed-origin": false,
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "403": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "404": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "408": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "409": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "413": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "422": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          },
          "503": {
            "description": "Product error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkHttpError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateTask": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "botId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "objective": {
            "type": "string",
            "minLength": 1,
            "maxLength": 16384
          },
          "tokenLimit": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000000
          },
          "requestKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NativeTaskScope"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "botId",
          "objective",
          "tokenLimit",
          "requestKey"
        ],
        "additionalProperties": false
      },
      "EmptyCommand": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {},
        "additionalProperties": false
      },
...
      "RequestCorrection": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "runId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "instruction": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4096
          },
          "requestKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "expectedSequence": {
            "type": "integer",
            "minimum": 0,
            "maximum": 8
          }
        },
        "required": [
          "runId",
          "instruction",
          "requestKey",
          "expectedSequence"
        ],
        "additionalProperties": false
      },
...
      "WorkHttpError": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
    }
  }
}